+++ src/binMeans.c
@@ -23,14 +23,14 @@ SEXP binMeans(SEXP y, SEXP x, SEXP bx, SEXP retCount, SEXP right) {
   assertArgVector(x, (R_TYPE_REAL), "x");
   nx = xlength(x);
   if (nx != ny) {
-    error("Argument 'y' and 'x' are of different lengths: %d != %d", ny, nx);
+    error("Argument 'y' and 'x' are of different lengths: %lld != %lld", (long long int)ny, (long long int)nx);
   }
 
   /* Argument 'bx': */
   assertArgVector(bx, (R_TYPE_REAL), "bx");
   nbins = xlength(bx)-1;
   if (nbins <= 0) {
-    error("Argument 'bx' must specify at least two bin boundaries (= one bin): %d", xlength(bx));
+    error("Argument 'bx' must specify at least two bin boundaries (= one bin): %lld", (long long int)xlength(bx));
   }
 
   /* Argument 'right': */