+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+ 
+
 # Version 1.1.0 [2023-11-06]
 
 ## Deprecated and Defunct