+++ NEWS.md
@@ -1,3 +1,15 @@
+# Version 1.2.0 [2023-12-11]
+
+## Bug Fixes
+
+ * Error messages that report on large integers (> 2^31 - 1), would
+   not render those integers correctly.
+
+## Deprecated and Defunct
+
+ * `useNames = NA` is defunct.
+
+
 # Version 1.1.0 [2023-11-06]
 
 ## Deprecated and Defunct